safiretools: centralizes shared Slater determinant utilities
